Bonus DONE Memory Leaks OK  Norminette OK

+++ libft/src/lib/ft_strchr.c
@@ -6,7 +6,7 @@
 /*   By: fporto <fporto@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/02/08 16:34:18 by fporto            #+#    #+#             */
-/*   Updated: 2021/10/21 03:57:51 by fporto           ###   ########.fr       */
+/*   Updated: 2022/03/09 16:29:27 by fporto           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 char	*ft_strchr(const char *s, int c)
 {
-	int	i;
+	size_t	i;
 
 	if (!s)
 		return (NULL);
